DjangoChallenge/manages/api/serializers.py
from rest_framework import serializers

from DjangoChallenge.manages.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'title', 'description', 'created_at', 'state']
